Moved ZIP archive handling into a reusable helper
ref https://linear.app/ghost/issue/MIG-1719

CSV ZIP imports need the same hardened extraction, filename, permission and structure rules as the /db/ importer without sharing ImportManager's mutable cleanup state. Isolating archive-only behavior lets both paths reuse those rules while keeping request cleanup independent.
